delete function issue fixed

+++ src/main/java/com/fightclub/controller/UserController.java
@@ -28,34 +28,36 @@ public class UserController {
 		return "user";
 	}
 
-	@RequestMapping(value = "/user.do", method = RequestMethod.POST)
-	public String doActions(@ModelAttribute @Valid User user, BindingResult result, @RequestParam String action,
+	@RequestMapping(value = "/user", method = RequestMethod.POST)
+	public String doActions(@Valid @ModelAttribute User user, BindingResult result, @RequestParam String action,
 			Map<String, Object> map) {
 		if (result.hasErrors()) {
-			return "/";
-		} else {
-			User userResult = new User();
-			switch (action.toLowerCase()) {
-			case "add":
-				userService.add(user);
-				userResult = user;
-				break;
-			case "edit":
-				userService.edit(user);
-				userResult = user;
-				break;
-			case "delete":
-				userService.delete(user.getId());
-				userResult = new User();
-				break;
-			case "search":
-				User searchedUser = userService.getUser(user.getId());
-				userResult = searchedUser != null ? searchedUser : new User();
-				break;
-			}
-			map.put("user", userResult);
-			map.put("userList", userService.getAllUsers());
-			return "user";
+			return "/user";
 		}
+		User userResult = new User();
+		switch (action.toLowerCase()) {
+		case "add":
+			userService.add(user);
+			userResult = user;
+			break;
+		case "edit":
+			userService.edit(user);
+			userResult = user;
+			break;
+		case "delete":
+			if (userService.getUser(user.getId()) != null) {
+				userService.delete(user.getId());
+			} else {
+				userResult = new User();
+			}
+			break;
+		case "search":
+			User searchedUser = userService.getUser(user.getId());
+			userResult = searchedUser != null ? searchedUser : new User();
+			break;
+		}
+		map.put("user", userResult);
+		map.put("userList", userService.getAllUsers());
+		return "user";
 	}
 }
